src/main/java/com/cgh/sell/service/impl/ProductCategoryServiceImpl.java
package com.cgh.sell.service.impl;

import com.cgh.sell.bean.ProductCategory;
import com.cgh.sell.dao.ProductCategoryDao;
import com.cgh.sell.service.ProductCategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class ProductCategoryServiceImpl implements ProductCategoryService {

    @Autowired
    private ProductCategoryDao productCategoryDao;

    @Override
    public ProductCategory findOne(Integer categoryId) {
        return productCategoryDao.findOne(categoryId);
    }

    @Override
    public List<ProductCategory> findAll() {
        return productCategoryDao.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> catrgoryTypeList) {
        return productCategoryDao.findByCategoryTypeIn(catrgoryTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return productCategoryDao.save(productCategory);
    }
}
